requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comet-ml==3.39.3
configobj==5.0.8
contourpy==1.1.1
cycler==0.12.1
docker-pycreds==0.4.0
dulwich==0.21.7
everett==3.1.0
filelock==3.13.3
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.6
importlib-resources==6.4.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning==2.2.1
lightning-utilities==0.11.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pillow==10.3.0
pkgutil-resolve-name==1.3.10
protobuf==4.25.3
psutil==5.9.8
pygments==2.17.2
pyparsing==3.1.2
python-box==6.1.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==1.44.1
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sympy==1.12
threadpoolctl==3.4.0
torch==2.2.2
torchmetrics==1.3.2
tqdm==4.66.2
typing-extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wandb==0.16.6
websocket-client==1.3.3
wrapt==1.16.0
wurlitzer==3.0.3
yarl==1.9.4
zipp==3.18.1
